RPi4: Add nic component

+++ mailboxInterface.c
@@ -6,9 +6,10 @@
 
 #include "mailboxInterface.h"
 
-#include <uspios.h>
-#include <uspi/bcm2835.h>
-#include <uspi/synchronize.h>
+#include <circleos.h>
+#include <circle/bcm2835.h>
+#include <circle/bcm2711.h>
+#include <circle/synchronize.h>
 
 #include <camkes.h>
 
@@ -30,8 +31,8 @@
 #define MAILBOX_STATUS_FULL     0x80000000
 
 /* Private function prototypes ----------------------------------------------------------------*/
-static uint32_t read32(uint32_t nAddress);
-static void write32(uint32_t nAddress, uint32_t nValue);
+static uint64_t read32(uint64_t nAddress);
+static void write32(uint64_t nAddress, uint64_t nValue);
 void MailboxFlush();
 unsigned MailboxRead(unsigned channel);
 void MailboxWrite(unsigned channel, unsigned nData);
@@ -43,7 +44,6 @@ bool MailboxInterface_getTag(uint32_t nTagId, void *pTag, unsigned nTagSize, uns
 	unsigned nBufferSize = sizeof (MailboxInterface_PropertyBuffer) + nTagSize + sizeof (uint32_t);
 
 	MailboxInterface_PropertyBuffer *pBuffer = (MailboxInterface_PropertyBuffer *)dma_alloc(DMA_PAGE_SIZE, DMA_ALIGNEMENT);
-
 	pBuffer->nBufferSize	= nBufferSize;
 	pBuffer->nCode 			= CODE_REQUEST;
 
@@ -57,9 +57,8 @@ bool MailboxInterface_getTag(uint32_t nTagId, void *pTag, unsigned nTagSize, uns
 
 	uint32_t *pEndTag 	= (uint32_t *) (pBuffer->Tags + nTagSize);
 	*pEndTag 			= PROPTAG_END;
-
 	uintptr_t physAddr 		= dma_getPhysicalAddr(pBuffer);
-
+	
 	uint32_t nBufferAddress = BUS_ADDRESS ((uint32_t) physAddr);
 
 	if (MailboxWriteRead (MAILBOX_CHANNEL, nBufferAddress) != nBufferAddress)
@@ -146,12 +145,12 @@ void MailboxFlush()
 	}
 }
 
-static uint32_t read32(uint32_t nAddress)
+static uint64_t read32(uint64_t nAddress)
 {
-	return *(volatile uint32_t *) nAddress;
+	return *(volatile uint64_t *) nAddress;
 }
 
-static void write32(uint32_t nAddress, uint32_t nValue)
+static void write32(uint64_t nAddress, uint64_t nValue)
 {
-	*(volatile uint32_t *) nAddress = nValue;
+	*(volatile uint64_t *) nAddress = nValue;
 }
